Move function helpers to their own package.
